Fix VirtualProtect() behaviour when setting copy-on-write protection
Existing Wine behaviour translates PAGE_WRITECOPY to PROT_WRITE | PROT_READ.
Presumably, this is because mprotect() cannot influence copy-on-write
behaviour for memory pages. Only mmap() can configure this. This results in
unexpected behaviour compared to Windows.

This change updates Wine's implementation of VirtualProtect() to detect when
copy-on-write protection is being set, and re-maps the memory as
copy-on-write using mmap(). This makes the behaviour consistent with what we
would expect to see on Windows.